[refactor] validation -> userInput
- validation파일의 유효성 검사를 userInput파일로 변경함
- 변경 후, 이 함수들 내에서 입력값을 받고 유효성 검사를 거쳐,
  App.js에서 사용될 수 있는 형태로 return하도록 함.
