+++ src/App.js
@@ -1,18 +1,15 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
 import { MESSAGES } from "./constants/message.js";
-import { isPlayerInputValid, isTrialInputValid } from "./utils/validation.js";
+import { playerInput, trialInput } from "./utils/userInput.js";
 import carsMove from "./utils/carsMove.js";
 import printResult from "./utils/printResult.js";
 
 class App {
   async play() {
     // 게임 시작
-    const playerInput = await MissionUtils.Console.readLineAsync(MESSAGES.START_INPUT);
-    const players = isPlayerInputValid(playerInput);
+    const players = await playerInput();
     const playersMap = new Map(players.map((player) => [player, 0]));
-
-    const tryInput = await MissionUtils.Console.readLineAsync(MESSAGES.TRY_INPUT);
-    const tryNum = isTrialInputValid(tryInput);
+    const tryNum = await trialInput();
 
     // 게임 진행
     MissionUtils.Console.print(MESSAGES.RESULT);
